R6: supersampling 2x, pase de tinta (multiply, micro-color, densidad intra-trazo, bleed)
Fase R6 del overhaul de realismo:
- supersample (default 2x): todo el render compone a Nx y reduce LANCZOS al
  final, POR PÁGINA (memoria); _scaled_options multiplica los px absolutos y
  el mínimo de cortesía de render_transparent escala por DPI (bug del wrapper:
  dejaba la página en 400/ss px).
- Refactor a pases (I2): capa de TINTA transparente + _compose_page →
  renderer_ink.apply_paper compone con MULTIPLY (la textura del papel
  atraviesa el trazo), operando solo el bbox con tinta.
- Micro-color HSV por glifo (D1, ink_hsv_jitter): la carga del bolígrafo
  nunca deposita el mismo color dos veces.
- Densidad intra-trazo (D2, ink_texture_strength): value noise (grid chico +
  BICUBIC, sin Perlin) modula el alpha DENTRO del trazo.
- Sangrado (D8, ink_bleed) como halo ADITIVO max(a, blur(a)·0.85): el blur
  directo despintaba el núcleo de los trazos finos y la tinta quedaba gris
  pálida (visto en golden); el halo expande el borde sin perder densidad.

Verificación: goldens con seed fija en tools/eval_render/golden/ (100/300%);
página completa R6 vs estilo-R5: 1.71x (< 2.5x del criterio); mismo seed →
bytes idénticos; suite completa 301 passed (se arreglaron de paso los 2
rojos de R0/R4 que dependían de este pase).

+++ core/inkcore/renderer_glyph.py
@@ -216,7 +216,7 @@ class GlyphLoadMixin:
                 if alpha_factor < 1.0:
                     a = a.point(lambda v: int(v * alpha_factor))
                 img = Image.merge("RGBA", (r, g, b, a))
-            return img, int(round(baseline_in)) if baseline_in >= 0 else -1
+            return img, round(baseline_in) if baseline_in >= 0 else -1
         except Exception as e:
             logger.debug(f"Could not load glyph {path}: {e}")
             return None
